+++ src/database/sqlite/sqlite_metadata.h
@@ -6,6 +6,24 @@
 #include "sqlite3.h"
 #include "sqlite_functions.h"
 
+// return a node list
+struct node_instance_list {
+    uuid_t  node_id;
+    uuid_t  host_id;
+    char *hostname;
+    int live;
+    int queryable;
+    int hops;
+};
+
+typedef enum db_check_action_type {
+    DB_CHECK_NONE          = (1 << 0),
+    DB_CHECK_RECLAIM_SPACE = (1 << 1),
+    DB_CHECK_ANALYZE       = (1 << 2),
+    DB_CHECK_CONT          = (1 << 3),
+    DB_CHECK_RECOVER       = (1 << 4),
+} db_check_action_type_t;
+
 // To initialize and shutdown
 void metadata_sync_init(void);
 void metadata_sync_shutdown(void);
@@ -19,6 +37,21 @@ void detect_machine_guid_change(uuid_t *host_uuid);
 void metadata_queue_load_host_context(RRDHOST *host);
 void metadata_delete_host_chart_labels(char *machine_guid);
 void vacuum_database(sqlite3 *database, const char *db_alias, int threshold, int vacuum_pc);
+
+int sql_metadata_cache_stats(int op);
+
+int get_node_id(uuid_t *host_id, uuid_t *node_id);
+int update_node_id(uuid_t *host_id, uuid_t *node_id);
+struct node_instance_list *get_node_list(void);
+void sql_load_node_id(RRDHOST *host);
+
+// Help build archived hosts in memory when agent starts
+void sql_build_host_system_info(uuid_t *host_id, struct rrdhost_system_info *system_info);
+void invalidate_node_instances(uuid_t *host_id, uuid_t *claim_id);
+RRDLABELS *sql_load_host_labels(uuid_t *host_id);
+
+int sql_init_meta_database(db_check_action_type_t rebuild, int memory);
+void sql_close_meta_database(void);
 
 // UNIT TEST
 int metadata_unittest(void);
